fix clang-tidy issue 1.2

+++ tasks/tabalaev_a_matrix_mul_strassen/seq/include/ops_seq.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stack>
 #include <vector>
 
 #include "tabalaev_a_matrix_mul_strassen/common/include/common.hpp"
@@ -28,13 +29,13 @@ class TabalaevAMatrixMulStrassenSEQ : public BaseTask {
   bool PostProcessingImpl() override;
 
   static std::vector<double> StrassenMultiply(const std::vector<double> &mat_a, const std::vector<double> &mat_b,
-                                              int n);
+                                              size_t n);
   static std::vector<double> Add(const std::vector<double> &mat_a, const std::vector<double> &mat_b);
   static std::vector<double> Subtract(const std::vector<double> &mat_a, const std::vector<double> &mat_b);
-  static std::vector<double> BaseMultiply(const std::vector<double> &mat_a, const std::vector<double> &mat_b, int n);
+  static std::vector<double> BaseMultiply(const std::vector<double> &mat_a, const std::vector<double> &mat_b, size_t n);
   static void PushStrassenSubtasks(std::stack<StrassenFrame> &frames, const std::vector<double> &mat_a,
-                                   const std::vector<double> &mat_b, int n);
-  static std::vector<double> CombineStrassenResults(std::stack<std::vector<double>> &results, int n);
+                                   const std::vector<double> &mat_b, size_t n);
+  static std::vector<double> CombineStrassenResults(std::stack<std::vector<double>> &results, size_t n);
 
   size_t a_rows_ = 0;
   size_t a_cols_b_rows_ = 0;
